chore: postIdx 속성 변경

+++ app/src/main/java/com/codepatissier/keki/src/main/consumer/store/storefeed/ConsumerStoreDetailFeedActivity.kt
@@ -111,7 +111,7 @@ class ConsumerStoreDetailFeedActivity : BaseActivity<ActivityConsumerStoreDetail
         })
     }
 
-    fun postLike(postIdx: Int){
+    fun postLike(postIdx: Long){
         ConsumerStoreFeedDetailService(this).tryPostConsumerStoreFeedDetailLike(postIdx)
     }
